(#130) Fix offset of CTextFile.OverridesTextMap
On b2372 a new slot for additional text was added which shifted some members of CTextFile, so offset 0x258 was the main text map instead of OverridesTextMap. This somehow caused the call to CTextFile.GetStringByHash to return NULL, at least in the example code.

Now the OverridesTextMap offset is found with a pattern at startup.