Fix keyboard layout pattern
Not finding this pattern caused the TokenParser hook to not localize keys correctly.

This was already broken before b2699 but was not noticed because the TokenParser has a fallback in case the keyboard layout is null.

Added some debug logs to avoid missing any other broken pattern in the future.